+++ internal/sdkprovider/service/clickhouse/clickhouse.go
@@ -12,6 +12,7 @@ func clickhouseSchema() map[string]*schema.Schema {
 	s[schemautil.ServiceTypeClickhouse] = &schema.Schema{
 		Type:        schema.TypeList,
 		Computed:    true,
+		Sensitive:   true,
 		Description: "Clickhouse server provided values",
 		Elem: &schema.Resource{
 			Schema: map[string]*schema.Schema{},