+++ internal/sdkprovider/service/dragonfly/dragonfly.go
@@ -12,6 +12,7 @@ func dragonflySchema() map[string]*schema.Schema {
 	s[schemautil.ServiceTypeDragonfly] = &schema.Schema{
 		Type:        schema.TypeList,
 		Computed:    true,
+		Sensitive:   true,
 		Description: "Dragonfly server provided values",
 		Elem: &schema.Resource{
 			Schema: map[string]*schema.Schema{},